refactor: ♻️ generalized popover component
